Ensure json data is re-read on every request
Prior to this change, it seemed that the json file
was being read once and then the data held in
memory.

This is more efficient, but it means that, when I
add a mechanism for updating the JSON data via an
http post, the new data would not be used
immediately.

If I want to return to in-memory caching, it 
should be easy enough to do that using a TTL for
the in-memory data.