Tweak the stylesheet tag
Without this change, the stylesheet was not 
applying correctly, when the app. runs inside a
docker container.

The error was something like:

Refused to apply style from '...' because its MIME type ('text/html') is not a supported stylesheet MIME type, and strict MIME checking is enabled.

I found the solution here:

https://stackoverflow.com/questions/48248832/stylesheet-not-loaded-because-of-mime-type#48270432