Allow variable number of data columns, in chart
This change adds a numeric column to the chart for
every column title provided after the first one,
which is assumed to be the namespace name.

This change will enable this template to be used
for the 'namespaces by pods' chart, which has one
fewer columns than the CPU & memory charts.